requirements.txt
discord.py[speed]
fastapi
uvicorn
aiohttp
python-dotenv
